Support using a custom network
Current implementation of the wallet is limited to the `mainnet` (called
`bitcoin`) and `testnet` networks. In order to allow the wallet to be
used in a different network it is required to define the new network
parameters and extend the code that relies on the "mainnet-vs-testnet"
logic to consider the possibility of a third custom network.

This change is mostly focused on extending the settings, BIP38 and
BIP44-related logic.